+++ src/components/App.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import Card from '@components/Card';
-
-const App = () => <div className='app' style={{ border: '1px solid black' }}><Card /></div>;
-
-export default App;