+++ src/components/App/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import Card from '@components/Card';
+import Style from './style';
+
+const App = () => <Style><div className='app'><Card /></div></Style>;
+
+export default App;